+++ classes/ShippingMethod.php
@@ -66,6 +66,7 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
         {
             $countries_options = $this->get_countries_options();
             $currency = get_woocommerce_currency();
+            $currency_symbol = get_woocommerce_currency_symbol();
             $fields = array(
                 'main_logo' => array(
                     'type' => 'logo'
@@ -214,6 +215,17 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
                 'description' => __('Send as one package', 'omniva_global'),
                 'type' => 'checkbox',
                 'default' => 'no',
+            );
+
+            $fields['add_delivery_time'] = array(
+                'title' => __('Add to delivery time', 'omniva_global'),
+                'description' => __('Add additional time in days to the delivery time shown to the customer', 'omniva_global'),
+                'type' => 'number',
+                'default' => 0,
+                'custom_attributes' => array(
+                    'step' => 1,
+                    'min' => 0
+                )
             );
             
             $services = $this->api->get_services();
@@ -486,6 +498,21 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
                 'default' => 2
             );
 
+            $fields['hr_advanced'] = array(
+                'type' => 'hr'
+            );
+
+            $fields['free_items_price'] = array(
+                'title' => __('Free items price', 'omniva_global'),
+                'type' => 'number',
+                'description' => sprintf(__('Some carriers do not allow the shipment to be registered if it contains items with a price of %1$s. In this setting, you can specify what price will be used when registering the shipment when the price of the item in the order is %1$s.', 'omniva_global'), '0 ' . $currency_symbol),
+                'custom_attributes' => array(
+                    'step' => 0.01,
+                    'min' => 0.01
+                ),
+                'default' => ''
+            );
+
             $fields['refresh_terminals'] = array(
                 'title' => __('Update terminals database', 'omniva_global'),
                 'type' => 'sync_button',
@@ -543,7 +570,7 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
             $service_key = $this->get_field_key( $key );
             $title = ($value['label'] ?? "");
             $html = '<tr valign="top"><th class="service-group-title">' . ucfirst($title) . '</th><td>';
-            $html .= '<label for="' . $service_key . '"><input type="checkbox" name="' . $service_key . '" id="' . $service_key . '" style="" value="yes" ' . ( $this->get_option( $key ) == 'yes' ? 'checked' : '' ) . '>' . __('Enable', 'omniva-global') . '</label>';
+            $html .= '<label for="' . $service_key . '"><input type="checkbox" name="' . $service_key . '" id="' . $service_key . '" style="" value="yes" ' . ( $this->get_option( $key ) == 'yes' ? 'checked' : '' ) . '>' . __('Enable', 'omniva_global') . '</label>';
             $html .= '</td></tr>';
             return $html;
         }
@@ -638,10 +665,14 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
                         }
                         $group = $offer->group;
                         $courier_title = $config[$group .'_title'] ?? 'Courier';
+                        $delivery_time = $this->get_delivery_time($offer->delivery_time);
+                        if (!empty($delivery_time)) {
+                            $delivery_time = ' (' . $delivery_time . ')';
+                        }
                         $free_shipping = $this->core->is_free_shipping($offer->group);
                         $rate = array(
                             'id' => $this->id . '_service_' . $offer->service_code,
-                            'label' => $courier_title . ' (' . $offer->delivery_time . ')',
+                            'label' => $courier_title . $delivery_time,
                             'cost' => $free_shipping ? 0 : $offer->price
                         );
                         $this->add_rate($rate);
@@ -672,6 +703,31 @@ if (!class_exists('\OmnivaTarptautinesWoo\ShippingMethod')) {
                 //}
             } catch (\Exception $e) {
             }
+        }
+
+        private function get_delivery_time($delivery_time)
+        {
+            $config = $this->core->get_config();
+
+            if (empty($delivery_time)) {
+                return '';
+            }
+
+            if (strpos($delivery_time, '-') === false) {
+                return $delivery_time;
+            }
+
+            $days = explode('-', $delivery_time);
+            $day_from = preg_replace("/[^0-9]/", "", $days[0]);
+            $day_to = preg_replace("/[^0-9]/", "", $days[1]);
+
+            if (!empty($config['add_delivery_time'])) {
+                $day_from += (int)$config['add_delivery_time'];
+                $day_to += (int)$config['add_delivery_time'];
+            }
+
+            return sprintf('%d-%d d.d.', $day_from, $day_to);
+            //return sprintf(_x('%d-%d days', 'Shipping time', 'omniva_global'), $day_from, $day_to); //Not working
         }
 
         public function process_admin_options()