+++ omniva-tarptautines-woo.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: Omniva International Shipping
- * Version: 1.1.0
+ * Version: 1.1.1
  * Plugin URI: https://github.com/mijora
  * Description: Official Omniva plugin that combine shipping between different countries
  * Author: Mijora
@@ -25,9 +25,9 @@ if (!defined('ABSPATH')) {
   exit;
 }
 
-define('OMNIVA_GLOBAL_VERSION', '1.1.0');
-define('OMNIVA_GLOBAL_BASENAME', plugin_basename(__FILE__));
-define('OMNIVA_GLOBAL_PLUGIN_DIR', plugin_dir_path( __FILE__ ));
+define('OMNIVALT_GLOBAL_VERSION', '1.1.1');
+define('OMNIVALT_GLOBAL_BASENAME', plugin_basename(__FILE__));
+define('OMNIVALT_GLOBAL_PLUGIN_DIR', plugin_dir_path( __FILE__ ));
 
 register_activation_hook(__FILE__, array( 'OmnivaTarptautinesWoo\Main', 'activated' ) );
 register_deactivation_hook( __FILE__, array( 'OmnivaTarptautinesWoo\Main', 'deactivated' ) );