+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.1.1]
+### Added
+- Added a ability to add additional days to delivery time
+- Added a ability to translate plugin
+- Added a ability to set item price in shipment, when item is free in order
+
+### Fixed
+- Fixed postcode when delivery address is in United Arab Emirates
+- Fixed items generation, when product name have HTML characters
+
 ## [1.1.0]
 ### Added
 - Added the option to select the price of the carrier to get with or without VAT